Implemented CreateGameModal
