+++ assets/js/pages/play/online/createGameModal.js
@@ -1,25 +1,12 @@
 import Modal from 'bootstrap/js/dist/modal.js';
-import onlinePlayersModal from './onlinePlayersModal.js';
-import ws from '../../../playWs.js';
-import * as mode from '../../../../mode.js';
+import CreateGameModal from '../../../components/CreateGameModal.js';
 
-const createGameModal = {
-  modal: new Modal(document.getElementById('createGameModal')),
-  form: document.querySelector('#createGameModal form')
-}
-
-createGameModal.form.addEventListener('submit', event => {
-  event.preventDefault();
-  const formData = new FormData(createGameModal.form);
-  const add = {
-    min: formData.get('minutes'),
-    increment: formData.get('increment'),
-    color: formData.get('color'),
-    submode: 'online'
-  };
-  localStorage.setItem('color', formData.get('color'));
-  ws.send(`/start ${formData.get('variant')} ${mode.PLAY} "${JSON.stringify(add).replace(/"/g, '\\"')}"`);
-  onlinePlayersModal.props.modal.show();
-});
+const createGameModal = new CreateGameModal(
+  document.getElementById('createGameModal'),
+  {
+    modal: new Modal(document.getElementById('createGameModal')),
+    form: document.querySelector('#createGameModal form')
+  }
+);
 
 export default createGameModal;
